Refactoring add gap bases so we can reuse some of the logic in taffy norm
